// FrontEnd/kausachunPeru/src/app/models/solicitudesCampesinos.ts
export class solicitudesCampesinos {
  id: String;
  detalle: String;
  comentario: String;
  fecha: String;
  idEstado: String;
  idTipo: String;
  idPersona: String;
  nombrePersona: String;
  apellidoPersona: String;
  dniPersona: String;
  tipoSolicitud: String;
  dptoPersona: String;
  estadoSolicitud: String;
}